requirements.txt
# Python >= 3.10 required
py-clob-client>=0.14.0
pyyaml>=6.0.0
requests>=2.31.0
aiohttp>=3.9.0

# Dashboard
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
